requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
scipy==1.11.1
requests==2.31.0
tqdm==4.65.0
